// lib/rickmorty/types.ts
// lib/rickmorty/types.ts
export type CharacterStatus = "Alive" | "Dead" | "unknown" | "Unknown";
export type CharacterGender = "Female" | "Male" | "Genderless" | "unknown" | "Unknown";


export type Character = {
  id: number;
  name: string;
  status: CharacterStatus;
  species: string;
  type: string;
  gender: CharacterGender;
  origin: { name: string; url: string };
  location: { name: string; url: string };
  image: string;
  episode: string[];
  url: string;
  created: string;
};

export type CharacterListResponse = {
  info: {
    count: number;
    pages: number;
    next: string | null;
    prev: string | null;
  };
  results: Character[];
};